feat(Engine): added layered renderable and animation

+++ src/R-Type/Systems/Bullets.system.cpp
@@ -29,7 +29,7 @@ void Bullets::tick()
     for (auto ent : WORLD.getEntitiesWithComponents<BaseBulletComponent, PositionComponent>()) {
         auto bullet = ent->getComponent<BaseBulletComponent>();
         auto pos    = ent->getComponent<PositionComponent>();
-        if (WINDOW.getSize().x + 100 <= pos->x || pos->x <= -100 || WINDOW.getSize().y + 100 <= pos->y ||
+        if ((int)WINDOW.getSize().x + 100 <= pos->x || pos->x <= -100 || (int)WINDOW.getSize().y + 100 <= pos->y ||
             pos->y <= -100) {
             WORLD.removeEntity(ent->getId());
         }
